feat: implement getting rate and subscribing

+++ internal/service/apis.go
@@ -22,5 +22,5 @@ type SendOptions struct {
 // CryptoAPI provides methods for getting crypto rates that are used in CryptoService and
 // implemented in external packages.
 type CryptoAPI interface {
-	GetRate(ctx context.Context, opts *GetRateOptions) (int, error)
+	GetRate(ctx context.Context, opts *GetRateOptions) (float64, error)
 }
